requirements.txt
PureCloudPlatformClientV2
python-dotenv